refactored /generate route into MVC model and wrote AI citation in readme
