+++ README.md
@@ -1,0 +1,4 @@
+AI Citation (Linh): 
+- JSDoc Documentation: used AI to generate API documentation for the /generate endpoint ; highlighted the code in cursor and asked it to generate a JSDoc description based on the given parameters / outputs 
+- Content generation prompts: used Cursor in-line prompting to create a prompt template for context-aware prompts that incorporates the project data and user requirements. Prompted along the lines of: "Based on the user-inputted data, generate a context-aware prompt that produces relevant marketing content. Incorporate the exact variables highlighted in the prompt"
+- AI integration: integrated with Google Cloud Vertex AI for actual content generation. Used GCP credits supplied in class.
